Add files via upload
This is a simple function that I will use in the future. Note it might have some bugs with russian but it works fine with bulgarian. The idea is that I will use it to translate post titles into link in latin alphabet and send the new text to be a link.